refactor: turn serializer and view in user from files into packages
