+++ apps/backend/users/views/__init__.py
@@ -1,0 +1,11 @@
+from .emails import EmailViewSet
+from .sessions import SessionViewSet
+from .subscriptions import UserSubscriptionViewSet
+from .users import UserViewSet
+
+__all__ = [
+    "EmailViewSet",
+    "SessionViewSet",
+    "UserSubscriptionViewSet",
+    "UserViewSet",
+]
